Move ConfirmedCasesByMunicipalitiesCard to components/index/cards
